@@ -13,9 +13,10 @@
   "dependencies": {
     "axios": "^0.24.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2"
+    "vue": "^3.2.41",
+    "@vue/compat": "^3.2.41",
+    "vue-router": "^4.1.6",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^1.6.1",
@@ -23,7 +24,8 @@
     "eslint-plugin-vue": "^9.6.0",
     "vite": "^2.5.4",
     "vite-plugin-vue2": "1.9.0",
-    "vue-cli-plugin-express": "~1.0.2"
+    "vue-cli-plugin-express": "~1.0.2",
+    "@vue/compiler-sfc": "^3.2.41"
   },
   "eslintConfig": {
     "root": true,
