@@ -1,18 +1,23 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
+import { createApp, h, configureCompat } from 'vue';
 import axios from 'axios';
 
 import App from './App.vue';
-import router from '@/router';
-import store from '@/store';
+import router from './router';
+import store from './store';
 
 import './assets/style/normalize.css';
 
-Vue.prototype.$axios = axios;
-Vue.use(VueRouter);
+const app = createApp({
+  render: () => h(App)
+});
+app.config.globalProperties.$axios = axios;
+app.use(router);
+app.use(store);
 
-new Vue({
-  router,
-  store,
-  render: h => h(App),
-}).$mount('#app')
+configureCompat({
+  // default everything to Vue 2 behavior
+  MODE: 2,
+  WATCH_ARRAY: false
+});
+
+app.mount('#app');
